Move utils dir into src
