+++ packages/contracts/src/utils/balances.ts
@@ -2,7 +2,7 @@ import { BigNumber } from '@0xproject/utils';
 import * as _ from 'lodash';
 import * as Web3 from 'web3';
 
-import { DummyTokenContract } from '../src/contract_wrappers/generated/dummy_token';
+import { DummyTokenContract } from '../contract_wrappers/generated/dummy_token';
 
 import { BalancesByOwner } from './types';
