+++ packages/contracts/src/utils/constants.ts
