+++ packages/contracts/src/utils/crypto.ts
