+++ packages/contracts/src/utils/exchange_wrapper.ts
@@ -3,7 +3,7 @@ import { BigNumber } from '@0xproject/utils';
 import * as _ from 'lodash';
 import * as Web3 from 'web3';
 
-import { ExchangeContract } from '../src/contract_wrappers/generated/exchange';
+import { ExchangeContract } from '../contract_wrappers/generated/exchange';
 
 import { formatters } from './formatters';
 import { signedOrderUtils } from './signed_order_utils';
