+++ packages/contracts/test/multi_sig_with_time_lock.ts
@@ -8,10 +8,10 @@ import * as Web3 from 'web3';
 import * as multiSigWalletJSON from '../../build/contracts/MultiSigWalletWithTimeLock.json';
 import { MultiSigWalletContract } from '../src/contract_wrappers/generated/multi_sig_wallet';
 import { MultiSigWalletWithTimeLockContract } from '../src/contract_wrappers/generated/multi_sig_wallet_with_time_lock';
-import { artifacts } from '../util/artifacts';
-import { constants } from '../util/constants';
-import { MultiSigWrapper } from '../util/multi_sig_wrapper';
-import { ContractName, SubmissionContractEventArgs } from '../util/types';
+import { artifacts } from '../src/utils/artifacts';
+import { constants } from '../src/utils/constants';
+import { MultiSigWrapper } from '../src/utils/multi_sig_wrapper';
+import { ContractName, SubmissionContractEventArgs } from '../src/utils/types';
 
 import { chaiSetup } from './utils/chai_setup';
 import { deployer } from './utils/deployer';
