+++ packages/contracts/test/multi_sig_with_time_lock_except_remove_auth_addr.ts
@@ -8,11 +8,11 @@ import * as Web3 from 'web3';
 import { MultiSigWalletContract } from '../src/contract_wrappers/generated/multi_sig_wallet';
 import { MultiSigWalletWithTimeLockExceptRemoveAuthorizedAddressContract } from '../src/contract_wrappers/generated/multi_sig_wallet_with_time_lock_except_remove_authorized_address';
 import { TokenTransferProxyContract } from '../src/contract_wrappers/generated/token_transfer_proxy';
-import { artifacts } from '../util/artifacts';
-import { constants } from '../util/constants';
-import { crypto } from '../util/crypto';
-import { MultiSigWrapper } from '../util/multi_sig_wrapper';
-import { ContractName, SubmissionContractEventArgs, TransactionDataParams } from '../util/types';
+import { artifacts } from '../src/utils/artifacts';
+import { constants } from '../src/utils/constants';
+import { crypto } from '../src/utils/crypto';
+import { MultiSigWrapper } from '../src/utils/multi_sig_wrapper';
+import { ContractName, SubmissionContractEventArgs, TransactionDataParams } from '../src/utils/types';
 
 import { chaiSetup } from './utils/chai_setup';
 import { deployer } from './utils/deployer';
