+++ packages/contracts/test/zrx_token.ts
@@ -6,8 +6,8 @@ import * as chai from 'chai';
 import * as Web3 from 'web3';
 
 import { ZRXTokenContract } from '../src/contract_wrappers/generated/zrx_token';
-import { constants } from '../util/constants';
-import { ContractName } from '../util/types';
+import { constants } from '../src/utils/constants';
+import { ContractName } from '../src/utils/types';
 
 import { chaiSetup } from './utils/chai_setup';
 import { deployer } from './utils/deployer';
